requirements.txt
# Core ML and Data Processing
tensorflow>=2.13.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# AWS Services
boto3>=1.28.0
sagemaker>=2.175.0
awscli>=1.29.0
aws-iot-device-sdk-python-v2>=1.15.0

# Data Processing & Streaming
paho-mqtt>=1.6.0
kafka-python>=2.0.2
apache-beam[gcp]>=2.48.0

# API and Web Services
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
websockets>=11.0.0

# Monitoring & Alerting
prometheus-client>=0.17.0
grafana-api>=1.0.3
python-dotenv>=1.0.0

# Data Validation & Quality
great-expectations>=0.17.0
pydantic>=2.0.0
marshmallow>=3.20.0

# Utilities
PyYAML>=6.0.0
click>=8.1.0
tqdm>=4.65.0
joblib>=1.3.0
python-dateutil>=2.8.0
pytz>=2023.3

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter & Notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Docker & Containerization
docker>=6.1.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0
mkdocs-material>=9.1.0
